+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KWBSJEV0G8B8E3YK9ZR72C18"
+        "01KWE79MPZ4KN9SKEW2VM71QR3"
     )
     r"""The unique identifier of the data source"""